Remove the explicit dependencies on notify gems
Guard wraps all of the notification stuff in some safe requires in older
versions (>= 2.3, < 2.11) and then extracts that behavior into Notiffany
after 2.11. This safe require behavior raises an error if the
appropriate library isn't installed, which -- to me at least -- is
acceptable behavior, since guard is a development tool anyway and won't
be run "in production".

The explicit requires cause havoc with using guard-rack in a vagrant
environment I'm working on. Just removing the explicit requires allows
me to run Guard in vagrant without issues. It also displays the
notifications properly to my notification daemon (I'm assuming via
libnotify, since I have it installed).

Do you agree that the message of "hey, you're missing a gem -- go
install it!" is an acceptable course of action?