Add support for a customizable rackup command
Instead of adding detection of a Gemfile and heuristics around when to
run `bundle exec rackup` vs. plain `rackup`, I figured it would be
better to go the way of guard-rspec and just allow the user to specify
the command that they want to run.

This also refactors the command builder into its own class to reduce the
complexity of the runner class and better encapsulate the logic.

Fixes #1