Use rack-compat to ensure we work with newer guard
Without this, we were having failing tests on guard versions >= 2.9.
That probably indicates a regression on their end, but since 2.9 was
released 5 months ago, it's probably not worth bringing up with them.

I verified that this change ensures compatibility for all guard versions
between 2.3 and 2.12.4, which is the entire scope of our ~> 2.3
dependency at this time.