Clamp Rack to < 2 on older Ruby versions
Rack 2 requires Ruby 2.2+. Since we test on 2.1, we need to clamp the
development dependency for Rack to < 2 when on the older Ruby.

I'm doing this programmatically instead of just doing it universally
because we *do* want to test on Rack 2 when we have a version of Ruby
that works with it. Since 2.1 is still in security-fix mode, it seems
like we should still test on it.